pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["Warren Snipes <contact@warrensnipes.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypdf = "5.0.0"
pillow = "10.4.0"
pydantic = "^2.9.2"
pymupdf = "^1.24.10"
nltk = "^3.9.1"
requests = "^2.32.3"
bs4 = "^0.0.2"
minio = "^7.2.10"
uuid = "^1.30"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
nbstripout = "^0.7.1"
pre-commit = "^4.0.1"
types-beautifulsoup4 = "^4.12.0.20240907"
types-requests = "^2.32.0.20241016"

[tool.poetry.group.llm-cuda.dependencies]
llama-cpp-python = { version = "^0.2.0", source = "llama-cpp-python" }
instructor = "^1.5.1"

[tool.poetry.group.llm.dependencies]
llama-cpp-python = { version = "^0.2.0" }
instructor = "^1.5.1"


[[tool.poetry.source]]
name = "llama-cpp-python"
url = "https://abetlen.github.io/llama-cpp-python/whl/cu122"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["nltk.*"]
ignore_missing_imports = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
